backend/requirements.txt
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==0.7.1
PyJWT==2.1.0
python-dotenv==0.19.1
Flask-Cors==3.10.9